# app/Models/User.php
<?php

namespace App\Models;

// use Illuminate\Contracts\Auth\MustVerifyEmail;
use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Foundation\Auth\User as Authenticatable;
use Illuminate\Notifications\Notifiable;
use Laravel\Sanctum\HasApiTokens;

class User extends Authenticatable
{
    /** @use HasFactory<\Database\Factories\UserFactory> */
    use HasFactory, Notifiable, HasApiTokens;

    /**
     * The attributes that are mass assignable.
     *
     * @var list<string>
     */
    protected $fillable = [
        'name',
        'email',
        'password',
        'role',
        'google_id',
        'email_verified_at'
    ];

    /**
     * The attributes that should be hidden for serialization.
     *
     * @var list<string>
     */
    protected $hidden = [
        'password',
        'remember_token',
    ];

    public function isAdmin() {
        return $this->role === 'admin';
    }
    
    public function isFormateur() {
        return $this->role === 'formateur';
    }

    public function lessonProgress()
    {
        return $this->hasMany(LessonUserProgress::class);
    }

    public function courseProgress()
    {
        return $this->hasMany(CourseUserProgress::class);
    }

    public function userQuizzes()
    {
        return $this->hasMany(UserQuiz::class);
    }

    public function completedLessons()
{
    return $this->hasMany(LessonUserProgress::class)
                ->where('is_completed', true);
}

    /**
     * Get the attributes that should be cast.
     *
     * @return array<string, string>
     */
    protected function casts(): array
    {
        return [
            'email_verified_at' => 'datetime',
            'password' => 'hashed',
        ];
    }
}