# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plateforme E-learning",
        "description": "Documentation API pour la plateforme e-learning",
        "contact": {
            "email": "diopawa425@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API Local"
        }
    ],
    "paths": {
...
        "/api/certificates/{id}": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Afficher un certificat",
                "operationId": "191e2a30f6ace09ac91b4b07289cfa86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du certificat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certificat PDF retourné",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "404": {
                        "description": "Certificat introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Lister tous les cours",
                "operationId": "02eeda67042dba2d22a517f26fb8eaf1",
                "responses": {
                    "200": {
                        "description": "Liste des cours",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Course"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Créer un nouveau cours",
                "operationId": "a4d6b7e7be59571034428ffbb86aaf0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Course"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cours créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get course details",
                "operationId": "f170bd35e3106c9d9bf5807499c80b68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/courses/{course}": {
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Modifier ou mettre a jour les inlfos du cours",
                "operationId": "5f1701987d5321072267c5f888fd05af",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Course"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Modifier avec succes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cours non trougvé"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/courses/{id}/start": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Démarrer le cours pour debloquer la première leçon.",
                "operationId": "a5923a7bbf1c26639f13739c687fd8fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cours démarré. Première leçon débloquée.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Content": {
                "required": [
                    "id",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "lesson_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "type": {
                        "type": "string",
                        "example": "pdf"
                    },
                    "data": {
                        "type": "string",
                        "nullable": true
                    },
                    "file_path": {
                        "type": "string",
                        "example": "lessons/files/example.pdf"
                    },
                    "external_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Course": {
                "title": "Course",
                "required": [
                    "id",
                    "title",
                    "description",
                    "progression",
                    "category",
                    "level",
                    "duration_minutes",
                    "is_published",
                    "user"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Initiation de base a l'informatique"
                    },
                    "description": {
                        "type": "string",
                        "example": "Learn the basics of Informatique"
                    },
                    "prerequis": {
                        "type": "string",
                        "nullable": true
                    },
                    "objectif": {
                        "type": "string",
                        "nullable": true
                    },
                    "progression": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "category": {
                        "type": "string",
                        "example": "Word"
                    },
                    "level": {
                        "type": "string",
                        "example": "Beginner"
                    },
                    "image_path": {
                        "type": "string",
                        "nullable": true
                    },
                    "duration_minutes": {
                        "type": "integer",
                        "example": 160
                    },
                    "is_published": {
                        "type": "boolean",
                        "example": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "lessons": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Lesson"
                        }
                    },
                    "resources": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "evaluations": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Lesson": {
                "required": [
                    "id",
                    "title",
                    "order",
                    "duration_minutes"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduction à Laravel"
                    },
                    "course_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    },
                    "duration_minutes": {
                        "type": "integer",
                        "example": 60
                    },
                    "is_published": {
                        "type": "boolean",
                        "example": false
                    },
                    "is_locked": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "contents": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Content"
                        }
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Awa Diop"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "diopawa192@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "passer"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Entrez le token Sanctum reçu après connexion.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}